Fix newParseException
