+++ hudi-spark-datasource/hudi-spark4.0.x/src/main/scala/org/apache/spark/sql/adapter/Spark4_0Adapter.scala
@@ -87,6 +87,13 @@ class Spark4_0Adapter extends BaseSpark4Adapter {
 
   override def getSparkPartitionedFileUtils: HoodieSparkPartitionedFileUtils = HoodieSpark40PartitionedFileUtils
 
+  override def newParseException(command: Option[String],
+                                 exception: AnalysisException,
+                                 start: Origin,
+                                 stop: Origin): ParseException = {
+    new ParseException(command, start, stop, exception.getErrorClass, exception.getMessageParameters.asScala.toMap)
+  }
+
   override def createAvroSerializer(rootCatalystType: DataType, rootType: HoodieSchema, nullable: Boolean): HoodieAvroSerializer =
     new HoodieSpark4_0AvroSerializer(rootCatalystType, rootType.toAvroSchema, nullable)
